Additional argument in Wallet CLI to pass already normalized tx amount

+++ wallet/src/main/scala/org/constellation/wallet/Transaction.scala
@@ -58,11 +58,10 @@ object Transaction {
     prevTx: Option[Transaction] = None,
     src: String,
     dst: String,
-    amount: Double,
+    amount: Long,
     keyPair: KeyPair,
     fee: Option[Double] = None
   ): Transaction = {
-
     val lastTxRef =
       prevTx
         .map(tx => LastTransactionRef(tx.hash, tx.lastTxRef.ordinal + 1))
